+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Extensible formatters for table columns in content element (#10)
 - Select alignment for table columns in content element (#11)
 - PSR-14 event to adjust or reject a dataset on synchronisation (#13)
+- Option in sync command for forcing a synchronisation
 
 ### Fixed
 - Domain model Table::setHandle() sets the correct property