+++ Documentation/changelog.rst
@@ -18,6 +18,7 @@ Added
 * Extensible formatters for table columns in content element (#10)
 * Select alignment for table columns in content element (#11)
 * PSR-14 event to adjust or reject a dataset on synchronisation (#13)
+* Option in sync command for forcing a synchronisation
 
 Fixed
 ^^^^^